+++ internal/provider/model_fmc_application_business_relevances.go
@@ -215,7 +215,7 @@ func (data *ApplicationBusinessRelevances) Clone() ApplicationBusinessRelevances
 func (data ApplicationBusinessRelevances) toBodyNonBulk(ctx context.Context, state ApplicationBusinessRelevances) string {
 	// This is one-by-one update, so only one element to update is expected
 	if len(data.Items) > 1 {
-		tflog.Error(ctx, "Found more than one element to chage. Only one will be changed.")
+		tflog.Error(ctx, "Found more than one element to change. Only one will be changed.")
 	}
 
 	// Utilize existing toBody function