+++ internal/provider/model_fmc_domains.go
@@ -218,7 +218,7 @@ func (data *Domains) Clone() Domains {
 func (data Domains) toBodyNonBulk(ctx context.Context, state Domains) string {
 	// This is one-by-one update, so only one element to update is expected
 	if len(data.Items) > 1 {
-		tflog.Error(ctx, "Found more than one element to chage. Only one will be changed.")
+		tflog.Error(ctx, "Found more than one element to change. Only one will be changed.")
 	}
 
 	// Utilize existing toBody function