+++ internal/provider/resource_fmc_security_zones_test.go
@@ -22,7 +22,10 @@ import (
 	"os"
 	"testing"
 
+	"github.com/hashicorp/terraform-plugin-testing/compare"
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
+	"github.com/hashicorp/terraform-plugin-testing/statecheck"
+	"github.com/hashicorp/terraform-plugin-testing/tfjsonpath"
 )
 
 // End of section. //template:end imports
@@ -84,3 +87,78 @@ func testAccFmcSecurityZonesConfig_all() string {
 }
 
 // End of section. //template:end testAccConfigAll
+
+func TestAccFmcSecurityZones_Sequential(t *testing.T) {
+
+	step_01 := `resource "fmc_security_zones" "test" {` + "\n" +
+		`	items = {` + "\n" +
+		`		"my_security_zones_1" = {` + "\n" +
+		`			interface_type = "ROUTED",` + "\n" +
+		`		},` + "\n" +
+		`		"my_security_zones_2" = {` + "\n" +
+		`			interface_type = "ROUTED",` + "\n" +
+		`		},` + "\n" +
+		`		"my_security_zones_3" = {` + "\n" +
+		`			interface_type = "ROUTED",` + "\n" +
+		`		},` + "\n" +
+		`	} ` + "\n" +
+		`}` + "\n"
+
+	var checks_step01 []resource.TestCheckFunc
+	checks_step01 = append(checks_step01, resource.TestCheckResourceAttrSet("fmc_security_zones.test", "items.my_security_zones_1.id"))
+	checks_step01 = append(checks_step01, resource.TestCheckResourceAttr("fmc_security_zones.test", "items.my_security_zones_1.interface_type", "ROUTED"))
+	checks_step01 = append(checks_step01, resource.TestCheckResourceAttrSet("fmc_security_zones.test", "items.my_security_zones_2.id"))
+	checks_step01 = append(checks_step01, resource.TestCheckResourceAttr("fmc_security_zones.test", "items.my_security_zones_2.interface_type", "ROUTED"))
+	checks_step01 = append(checks_step01, resource.TestCheckResourceAttrSet("fmc_security_zones.test", "items.my_security_zones_3.id"))
+	checks_step01 = append(checks_step01, resource.TestCheckResourceAttr("fmc_security_zones.test", "items.my_security_zones_3.interface_type", "ROUTED"))
+
+	step_02 := `resource "fmc_security_zones" "test" {` + "\n" +
+		`	items = {` + "\n" +
+		`		"my_security_zones_1" = {` + "\n" +
+		`			interface_type = "ROUTED",` + "\n" +
+		`		},` + "\n" +
+		`		"my_security_zones_2" = {` + "\n" +
+		`			interface_type = "INLINE",` + "\n" +
+		`		},` + "\n" +
+		`		"my_security_zones_4" = {` + "\n" +
+		`			interface_type = "ROUTED",` + "\n" +
+		`		},` + "\n" +
+		`	} ` + "\n" +
+		`}` + "\n"
+
+	var checks_step02 []resource.TestCheckFunc
+	checks_step02 = append(checks_step02, resource.TestCheckResourceAttrSet("fmc_security_zones.test", "items.my_security_zones_1.id"))
+	checks_step02 = append(checks_step02, resource.TestCheckResourceAttr("fmc_security_zones.test", "items.my_security_zones_1.interface_type", "ROUTED"))
+	checks_step02 = append(checks_step02, resource.TestCheckResourceAttrSet("fmc_security_zones.test", "items.my_security_zones_2.id"))
+	checks_step02 = append(checks_step02, resource.TestCheckResourceAttr("fmc_security_zones.test", "items.my_security_zones_2.interface_type", "INLINE"))
+	checks_step02 = append(checks_step02, resource.TestCheckNoResourceAttr("fmc_security_zones.test", "items.my_security_zones_3.id"))
+	checks_step02 = append(checks_step02, resource.TestCheckResourceAttrSet("fmc_security_zones.test", "items.my_security_zones_4.id"))
+	checks_step02 = append(checks_step02, resource.TestCheckResourceAttr("fmc_security_zones.test", "items.my_security_zones_4.interface_type", "ROUTED"))
+
+	// Zone 1 should not change
+	zone1Id := statecheck.CompareValue(compare.ValuesSame())
+	// Zone 2 should be recreated, due to the change in interface_type
+	zone2Id := statecheck.CompareValue(compare.ValuesDiffer())
+
+	steps := []resource.TestStep{{
+		Config: step_01,
+		Check:  resource.ComposeTestCheckFunc(checks_step01...),
+		ConfigStateChecks: []statecheck.StateCheck{
+			zone1Id.AddStateValue("fmc_security_zones.test", tfjsonpath.New("items").AtMapKey("my_security_zones_1").AtMapKey("id")),
+			zone2Id.AddStateValue("fmc_security_zones.test", tfjsonpath.New("items").AtMapKey("my_security_zones_2").AtMapKey("id")),
+		},
+	}, {
+		Config: step_02,
+		Check:  resource.ComposeTestCheckFunc(checks_step02...),
+		ConfigStateChecks: []statecheck.StateCheck{
+			zone1Id.AddStateValue("fmc_security_zones.test", tfjsonpath.New("items").AtMapKey("my_security_zones_1").AtMapKey("id")),
+			zone2Id.AddStateValue("fmc_security_zones.test", tfjsonpath.New("items").AtMapKey("my_security_zones_2").AtMapKey("id")),
+		},
+	}}
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { testAccPreCheck(t) },
+		ProtoV6ProviderFactories: testAccProtoV6ProviderFactories,
+		Steps:                    steps,
+	})
+}